requirements.txt
# Project Raseed - FastAPI Backend Dependencies

aiohttp
# Core FastAPI and web framework
fastapi

# Google Agent Development Kit (ADK)
google-adk

# Vertex AI specific dependencies
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform

# Google Cloud services
google-cloud-firestore
google-cloud-storage
google-cloud-vision
google-generativeai

# Video processing (optional, for creating test videos)
# Install with: pip install opencv-python
# Used in test_video_receipt.py for creating synthetic videos

# Production server
gunicorn

# HTTP and async support
httpx
langchain-google-vertexai
opencv-python

# Image and Video processing
Pillow
psutil
pydantic
# Pillow==10.4.0
# pydantic==2.8.2
pydantic-settings

# Development and testing
pytest
pytest-asyncio
python-dotenv
python-multipart
requests
shortuuid

# Utilities
typing-extensions
uvicorn
